+++ Locales/enUS.lua
@@ -25,6 +25,10 @@ local L = Shared.Locale
 --L["Price"] = "Price"
 --L["Usable"] = "Usable"
 --L["Available"] = "Available"
+--L["Index"] = "Index"
+--L["Item ID"] = "Item ID"
+--L["Type"] = "Type"
+--L["Subtype"] = "Subtype"
 
 -- Using short keys for these long strings, so enUS needs to be defined as well
 L["OPTIONS_TITLE_MAIN"] = "%s Main Options"
@@ -32,3 +36,5 @@ L["OPTIONS_TITLE_TAB_DEFAULT"] = "Show %s tab by default"
 L["OPTIONS_DESCRIPTION_TAB_DEFAULT"] = "Use the enhanced merchant interface for all merchants. The standard merchant tab can still be accessed at any time."
 L["OPTIONS_TITLE_SORT_REMEMBER"] = "Remember sort order"
 L["OPTIONS_DESCRIPTION_SORT_REMEMBER"] = "Keep the last used sort when switching between merchants. You can reset a merchant to the game's standard sort by holding CTRL while clicking on a column heading."
+L["OPTIONS_TITLE_COLUMNS"] = "Displayed Columns"
+L["OPTIONS_DESCRIPTION_COLUMNS"] = "Select the columns to be shown on merchants."